+++ downloads.sh
@@ -2,11 +2,11 @@
 
 ## SOFTWARE: fastqc
 ## VERSION: 0.11.5
-wget https://www.bioinformatics.babraham.ac.uk/projects/fastqc/fastqc_v0.11.5.zip
-unzip fastqc_v0.11.5.zip
-cd FastQC/
-chmod 755 fastqc
-cd ..
+#wget https://www.bioinformatics.babraham.ac.uk/projects/fastqc/fastqc_v0.11.5.zip
+#unzip fastqc_v0.11.5.zip
+#cd FastQC/
+#chmod 755 fastqc
+#cd ..
 
 
 ## SOFTWARE: bwa